requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.2
django-cors-headers==4.5.0
django-extensions==3.2.3
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
MarkupSafe==3.0.2
phonenumbers==8.13.48
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
sqlparse==0.5.1
tzdata==2024.2
Werkzeug==3.1.3
